Add address->uuid mapping table to partition state
When a member is restarted in a short period, before it's removed
from partition table in time, we need to identify that an old member is
removed and started back as a new member. This is required to be able
to repair partition table correctly.
A mapping for address->uuid is kept alongside with partition
table and during repair we lookup a member's uuid in that mapping
table. If it's not found, it means that member is a new member.