Fix a membership change bug in the Raft core and server modules
Consider the following scenario:
- There are 3 members in the raft group: [A, B, C].
- We will remove C, then add D.
- We first commit <remove C>, and the group becomes [A, B].
- We then start D and join it to the raft group. D's initial member list is [D].
- A and B successfully commits <add D> and their committed member list becomes [A, B, D].
- D receives the log from the leader and first commits <remove C> and updates its member list as [A, B].
- While applying <remove C>, D also checks if it is present in the current member list. Since it is not, it updates its status as STEPPED_DOWN.
- Now D is broken. Since it is STEPPED_DOWN, it will not append entries anymore.

There are one major and one minor issue in this scenario:
The major issue is, converting to STEPPING_DOWN should be done only by checking which member is removed via ApplyRaftGroupMembersCmd
The minor issue is, there is no point in starting a joined node with a singleton member list of itself. As we see in this scenario, even if D starts with the member list of [D], once it applies <remove C>, it already gets the member list of [A, B], in which D is not present. So it is totally valid when a joined member is not in its committed member list yet, although it is not removed. This can happen while a joined member commits the appended entries from the beginning, as in this scenario. It also breaks the deterministic status of the state machine.