Make fenced lock re-entrancy non-local
Consider the following scenario:

1) client.lock();
Successful.
2) The client encounters a network partition
3) client.unlock();
Fails with operation timeout and also not executed
on the server, but still local lock state is deleted.
4) Network partition issue is gone. The client still owns the lock.
5) client.getLockCount();
Returns 1, because client asks to the server.
6) client.lock();
Successful
7) client.getLockCount();
Returns 1, because client uses the local lock state to get the count,
but the actual lock count is 2.

Keeping local lock state is a difficult task and easily leads to
problematic scenarios as above.