Introduce InvocationTargetLeaveAware for raft operations
InvocationTargetLeaveAware is used for determining if a raft operation
is safe to retry on leader leave and it is ok to commit the operation
multiple times. If the operation is safe, it is retried automatically if
the leader returns commit response of the invoked operation. If the
operation is not known to be safe, i.e. it either does not implement
InvocationTargetLeaveAware or returns false from
InvocationTargetLeaveAware#isSafeToRetryOnTargetLeave(), then
retry is decided based on the
RaftAlgorithmConfig#failOnIndeterminateOperationState config parameter.

Safe to retry raft operations are expected to be idempotent.
For instance, lock, session-aware semaphore, and count down latch
operations are idempotent. It does not make a difference to commit an
idempotent operation once or multiple times. However, although some
raft operations are idempotent in terms of state changes they cause,
they may not return the same response on duplicate commits. They can be
still considered to be safe to retry on leader node leave, depending on
their context and how they are used.