Check target address is not local in NodeEngineRaftIntegration
When a Hazelcast member leaves a CP group, shuts down and restart,
it can join to the CP group with the same ip but a different CP member
uuid. When it initializes its local RaftNode, its previous CP member
identity will be present in the CP group member list. In this case,
it will consider its previous CP member identity as a remote endpoint.